chore(admin): Move tests directory
