+++ __tests__/admin/query.test.ts
@@ -1,4 +1,4 @@
-import { FirestoreSimpleAdmin } from '../src'
+import { FirestoreSimpleAdmin } from '../../src'
 import { createRandomCollectionName, deleteCollection, initFirestore } from './util'
 
 export interface TestDoc {
